Split Collator/Formatted types into separate files
This avoids a troublesome circular import.
