+++ src/style-spec/expression/definitions/format.js
@@ -1,6 +1,7 @@
 // @flow
 
 import { NumberType, ValueType, FormattedType, array, StringType } from '../types';
+import Formatted, { FormattedSection } from '../types/formatted';
 import { toString } from '../values';
 
 import type { Expression } from '../expression';
@@ -8,49 +9,13 @@ import type EvaluationContext from '../evaluation_context';
 import type ParsingContext from '../parsing_context';
 import type { Type } from '../types';
 
-export class FormattedSection {
-    text: string
-    scale: number | null
-    fontStack: string | null
-
-    constructor(text: string, scale: number | null, fontStack: string | null) {
-        this.text = text;
-        this.scale = scale;
-        this.fontStack = fontStack;
-    }
-}
-
-export class Formatted {
-    sections: Array<FormattedSection>
-
-    constructor(sections: Array<FormattedSection>) {
-        this.sections = sections;
-    }
-
-    toString(): string {
-        return this.sections.map(section => section.text).join('');
-    }
-
-    serialize() {
-        const serialized = ["format"];
-        for (const section of this.sections) {
-            serialized.push(section.text);
-            const fontStack = section.fontStack ?
-                ["literal", section.fontStack.split(',')] :
-                null;
-            serialized.push({ "text-font": fontStack, "font-scale": section.scale });
-        }
-        return serialized;
-    }
-}
-
 type FormattedSectionExpression = {
     text: Expression,
     scale: Expression | null;
     font: Expression | null;
 }
 
-export class FormatExpression implements Expression {
+export default class FormatExpression implements Expression {
     type: Type;
     sections: Array<FormattedSectionExpression>;
