+++ src/style-spec/expression/values.js
@@ -3,9 +3,9 @@
 import assert from 'assert';
 
 import Color from '../util/color';
-import { Collator } from './definitions/collator';
-import { Formatted } from './definitions/formatted';
-import { NullType, NumberType, StringType, BooleanType, ColorType, ObjectType, ValueType, CollatorType, array } from './types';
+import Collator from './types/collator';
+import Formatted from './types/formatted';
+import { NullType, NumberType, StringType, BooleanType, ColorType, ObjectType, ValueType, CollatorType, FormattedType, array } from './types';
 
 import type { Type } from './types';
 
@@ -28,7 +28,7 @@ export function validateRGBA(r: mixed, g: mixed, b: mixed, a?: mixed): ?string {
     return null;
 }
 
-export type Value = null | string | boolean | number | Color | Collator | $ReadOnlyArray<Value> | { +[string]: Value }
+export type Value = null | string | boolean | number | Color | Collator | Formatted | $ReadOnlyArray<Value> | { +[string]: Value }
 
 export function isValue(mixed: mixed): boolean {
     if (mixed === null) {
@@ -42,6 +42,8 @@ export function isValue(mixed: mixed): boolean {
     } else if (mixed instanceof Color) {
         return true;
     } else if (mixed instanceof Collator) {
+        return true;
+    } else if (mixed instanceof Formatted) {
         return true;
     } else if (Array.isArray(mixed)) {
         for (const item of mixed) {
@@ -75,6 +77,8 @@ export function typeOf(value: Value): Type {
         return ColorType;
     } else if (value instanceof Collator) {
         return CollatorType;
+    } else if (value instanceof Formatted) {
+        return FormattedType;
     } else if (Array.isArray(value)) {
         const length = value.length;
         let itemType: ?Type;
