+++ src/style/style_layer/fill_style_layer_properties.js
@@ -15,7 +15,7 @@ import {
 
 import type Color from '../../style-spec/util/color';
 
-import type {Formatted} from '../../style-spec/expression/definitions/formatted';
+import type Formatted from '../../style-spec/expression/types/formatted';
 
 
 export type PaintProps = {|
