+++ src/style/style_layer/layer_properties.js.ejs
@@ -20,7 +20,7 @@ import {
 
 import type Color from '../../style-spec/util/color';
 
-import type {Formatted} from '../../style-spec/expression/definitions/formatted';
+import type Formatted from '../../style-spec/expression/types/formatted';
 
 <% if (layoutProperties.length) { -%>
 export type LayoutProps = {|
